LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-relevant-mutations-expected.txt

CONSOLE MESSAGE: Error: assert_false: below-viewport-1 image should never load expected false got true
   

Harness Error (FAIL), message = Error: assert_false: below-viewport-1 image should never load expected false got true

TIMEOUT Images are lazyloaded Test timed out
NOTRUN Image referrerPolicy mutation does not cause deferred loading=lazy images to be fetched 
NOTRUN Image crossOrigin mutation does not cause deferred loading=lazy images to be fetched 
NOTRUN Image src mutation does not cause deferred loading=lazy images to be fetched 
